feat(auth): improve oauth browser flow
Centralizes the browser detection logic into a shared utility and applies it consistently across the application. This allows the tool to automatically detect non-interactive environments (like
SSH sessions or CI/CD) and default to a manual, code-based authentication flow.

If a browser launch fails in an interactive environment, the application now advises the user to restart with NO_BROWSER=true, which is the correct procedure for the authcode flow. This
replaces a flawed inline fallback.
