+++ packages/core/src/index.ts
@@ -77,3 +77,4 @@ export { OAuthUtils } from './mcp/oauth-utils.js';
 // Export telemetry functions
 export * from './telemetry/index.js';
 export { sessionId } from './utils/session.js';
+export * from './utils/browser.js';
